Added a way for contacts to rent books
created a way to rent books for contacts (res.partner), created tree, form and calendar view (for user friendlyness simplification) added state which is not predefined, but calculated based on current day and start of rent. if the start of rent is not today its a reservation if its today its rented if in form cancel is pressed and it was a reservation it is canceled if return is pressed the client has returned the book and the book was rented, not reserved. started on scheduled checking for late fees